+++ .gitignore
@@ -35,6 +35,7 @@ jspm_packages
 .node_repl_history
 
 dist
+build
 bin
 cdn
 api-javascript
