+++ index.js
@@ -1,9 +1,1 @@
-(function (w) {
-  if (typeof module !== 'undefined' && typeof module.exports !== 'undefined') {
-    module.exports = require('./bin/gb-tracker-client');
-  }
-
-  if (w) {
-    w.GbTracker = require('./bin/gb-tracker-client');
-  }
-})(window);
+module.exports = require('./build/gb-tracker-full');
