+++ lib/gb-tracker-window.js
@@ -1,7 +1,7 @@
 (function (w) {
   if (typeof module !== 'undefined' && typeof module.exports !== 'undefined') {
-    module.exports = require('./gb-tracker-core');
+    module.exports = require('./gb-tracker-full');
   }
 
-  w.GbTracker = require('./gb-tracker-core');
+  w.GbTracker = require('./gb-tracker-full');
 })(window);
